python_recv_esp32/requirements.txt
mediapipe==0.8.9.1
numpy==1.21.6
opencv_python==4.5.3.56
paho_mqtt==1.6.1
requests==2.26.0
